Timestamp is u32

+++ src/utils/mod.rs
@@ -4,11 +4,11 @@ pub use decode::*;
 
 mod decode;
 
-pub fn local_timestamp() -> u64 {
+pub fn local_timestamp() -> u32 {
     SystemTime::now()
         .duration_since(UNIX_EPOCH)
         .expect("Time went backwards")
-        .as_secs()
+        .as_secs() as u32
 }
 
 pub fn median<T: Clone>(inps: &Vec<T>) -> T {
